+++ src/clj/spring_lobby/fs/sdfz.clj
@@ -42,13 +42,67 @@
 (def sdfz-protocol
   (b/header
     sdfz-header
-    (fn [{:keys [script-size]}]
+    (fn [{:keys [demo-stream-size script-size]}]
       (b/ordered-map
         :something :int-le
         :something2 :int-le
-        :script-txt (b/string "ISO-8859-1" :length script-size)))
+        :script-txt (b/string "ISO-8859-1" :length script-size)
+        :demo-stream (b/blob :length demo-stream-size)))
     (constantly nil) ; TODO writing replays
     :keep-header? true))
+
+
+; https://github.com/spring/spring/blob/develop/rts/Net/Protocol/NetMessageTypes.h
+(def net-message-header
+  (b/ordered-map
+    :command :byte))
+
+(defn net-message-protocol [length]
+  (b/header
+    net-message-header
+    (fn [{:keys [command]}]
+      (let [l (dec length)]
+        (case command
+          6 (b/ordered-map
+              :pad :byte
+              :player-num :ubyte
+              :player-name (b/string "ISO-8859-1" :length (- length 3)))
+          7 (b/ordered-map
+              :pad :byte
+              :from :ubyte
+              :dest :ubyte
+              :message (b/string "ISO-8859-1" :length (- length 4)))
+          30 (b/ordered-map
+               :pad :byte
+               :player-num :ubyte
+               :winning-ally-teams (b/repeated :ubyte :length (- length 3)))
+          36 (b/ordered-map
+               :player-num :ubyte
+               :my-team :ubyte
+               :ready :ubyte
+               :x :float-le
+               :y :float-le
+               :z :float-le)
+          (b/blob :length l))))
+    (constantly nil) ; TODO writing replays
+    :keep-header? true))
+
+
+; https://github.com/beyond-all-reason/spring/blob/da07d8db67269bf74fb68856431c1a18e7d66600/rts/System/LoadSave/demofile.h#L98-L103
+(def demo-stream-chunk-header
+  (b/ordered-map
+    :mod-game-time :float-le
+    :length :uint-le))
+
+(def demo-stream-chunk-protocol
+  (b/header
+    demo-stream-chunk-header
+    (fn [{:keys [length]}]
+      (b/ordered-map
+        :demo-stream-chunk (net-message-protocol length)))
+    (constantly nil) ; TODO writing replays
+    :keep-header? true))
+
 
 (defn decode-replay-header [^java.io.File f]
   (with-open [is (io/input-stream f)
@@ -60,21 +114,32 @@
 (defn sanitize-replay-engine [replay-engine-version]
   (some-> replay-engine-version (string/replace #"\P{Print}" "") string/trim fs/sync-version-to-engine-version))
 
-(defn decode-replay [^java.io.File f]
-  (with-open [is (io/input-stream f)
-              gz (GZIPInputStream. is)]
-    (-> (->> (b/decode sdfz-protocol gz)
-             (into (sorted-map)))
-        (update :header
-          (fn [header]
-            (-> header
-                (update :engine-version sanitize-replay-engine)
-                (update :game-id u/bytes->hex))))
-        (update :body
-          (fn [body]
-            (-> body
-                (assoc :script-data (spring-script/parse-script (:script-txt body)))
-                (dissoc :script-txt)))))))
+(defn decode-replay
+  ([^java.io.File f]
+   (decode-replay f nil))
+  ([^java.io.File f {:keys [parse-stream]}]
+   (with-open [is (io/input-stream f)
+               gz (GZIPInputStream. is)]
+     (-> (->> (b/decode sdfz-protocol gz)
+              (into (sorted-map)))
+         (update :header
+           (fn [header]
+             (-> header
+                 (update :engine-version sanitize-replay-engine)
+                 (update :game-id u/bytes->hex))))
+         (update :body
+           (fn [body]
+             (-> body
+                 (assoc :script-data (spring-script/parse-script (:script-txt body)))
+                 (dissoc :script-txt)
+                 (update :demo-stream
+                   (fn [demo-stream-raw]
+                     (if parse-stream
+                       (try
+                         (b/decode (b/repeated demo-stream-chunk-protocol) (io/input-stream demo-stream-raw))
+                         (catch Exception e
+                           (log/warn e "Exception parsing demo stream")))
+                       nil))))))))))
 
 
 (defn replay-game-type [allyteam-counts]
@@ -104,14 +169,17 @@
     {:game-type (replay-game-type allyteam-counts)
      :player-counts allyteam-counts}))
 
-(defn parse-replay [^java.io.File f]
-  (let [replay (try
-                 (decode-replay f)
-                 (catch Exception e
-                   (log/error e "Error reading replay" f)))]
-    (merge
-      {:file f
-       :filename (fs/filename f)
-       :file-size (fs/size f)}
-      replay
-      (replay-type-and-players replay))))
+(defn parse-replay
+  ([^java.io.File f]
+   (parse-replay f nil))
+  ([^java.io.File f opts]
+   (let [replay (try
+                  (decode-replay f opts)
+                  (catch Exception e
+                    (log/error e "Error reading replay" f)))]
+     (merge
+       {:file f
+        :filename (fs/filename f)
+        :file-size (fs/size f)}
+       replay
+       (replay-type-and-players replay)))))